visualization and prepare datasets v1
