+++ datasets/README.md
@@ -5,3 +5,9 @@
 - The wiki_test_case dataset is composed from the first Table found in: https://en.wikipedia.org/wiki/Lift_(data_mining)
 
 - The Breast Cancer dataset is downloaded from http://archive.ics.uci.edu/ml/datasets/Breast+Cancer
+
+- The weather_data dataset is downloaded from https://www.kaggle.com/datasets/prasad22/weather-data
+
+- The football_players dataset is downloaded from https://www.kaggle.com/datasets/berkayalan/the-most-expensive-football-transfers
+
+- The data_developer_salary dataset is downloaded from https://www.kaggle.com/datasets/zeesolver/data-eng-salary-2024
